Make sure streams are piping to containres
Ok, this is weird but here's my explanation. We start out by making a POST request with the filename to get a UUID. We do that so that we can get a specific place to start streaming the file to without waiting for it to upload entirely. In fact, I found that the file wouldn't upload because it was piped to a stream that wasn't consumed. This way, we have a place to start uploading the file that we can use for constructing the command to the `/run/image` endpoint. That way, the running container will consume the file stream and it will upload straight into the container.

I only really checked out the memory consumption with activity monitor.app and it looked like a 104MB file added about 10MB of RAM usage to the app but it streamed to the container just fine.